move dynamic to its own pkg (1)
